+++ Changelog.md
@@ -3,6 +3,7 @@ Changelog
 
 ### wunderbar 1.0 b4 (not released yet!)
 * [Feature] Sort tasks by due date
+* [Feature] wunderbar displays a traybubble if the user openend the application the first time.
 
 ### wunderbar 1.0 b3
 * [Bug] In some cases the transmitted date-value can be null which caused an NullReferenceException.